relocated screens into separate folders
