+++ screens/room/RoomBooking.js
@@ -3,14 +3,14 @@ import { useState, useEffect } from 'react';
 import { useIsFocused } from '@react-navigation/native';
 import Moment from 'moment';
 
-import Input from '../components/Input';
-import StandardButton from '../components/buttons/StandardButton';
-import BackButton from '../components/buttons/BackButton';
+import Input from '../../components/Input';
+import StandardButton from '../../components/buttons/StandardButton';
+import BackButton from '../../components/buttons/BackButton';
 
-import textStyle from '../styles/text';
-import colors from '../styles/colors';
+import textStyle from '../../styles/text';
+import colors from '../../styles/colors';
 
-import { loadSecure } from '../utils/secureStore';
+import { loadSecure } from '../../utils/secureStore';
 
 
 export default function RoomBooking({ navigation, route }) {
